Finalizando refatoração nivel
