+++ app/Helpers/DashboardHelper.php
@@ -4,11 +4,11 @@ namespace App\Helpers;
 
 use App\Models\Parametro;
 use App\Services\AdministradorService;
-use App\Services\DiretoriaService;
+use App\Services\Instancias\DiretoriaService;
 use App\Services\Estatistica\EstatisticaService;
-use App\Services\FederacaoService;
-use App\Services\LocalService;
-use App\Services\SinodalService;
+use App\Services\Instancias\FederacaoService;
+use App\Services\Instancias\LocalService;
+use App\Services\Instancias\SinodalService;
 
 class DashboardHelper
 {
@@ -34,11 +34,6 @@ class DashboardHelper
     {
         $class = self::make();
         return $class::getTotalizadores();
-    }
-
-    public static function getTotalizadoresEstatisticaExterno()
-    {
-        return EstatisticaService::getTotalizadores();
     }
 
     public static function getInfo()
