+++ routes/web.php
@@ -12,16 +12,16 @@ use App\Http\Controllers\DemandaController;
 use App\Http\Controllers\DigestoController;
 use App\Http\Controllers\EstatisticaController;
 use App\Http\Controllers\Produtos\EstoqueProdutoController;
-use App\Http\Controllers\FederacaoController;
+use App\Http\Controllers\Instancias\FederacaoController;
 use App\Http\Controllers\Formularios\FormularioFederacaoController;
 use App\Http\Controllers\Formularios\FormularioLocalController;
 use App\Http\Controllers\Formularios\FormularioSinodalController;
 use App\Http\Controllers\HomeController;
-use App\Http\Controllers\LocalController;
+use App\Http\Controllers\Instancias\LocalController;
 use App\Http\Controllers\MinhasDemandasController;
 use App\Http\Controllers\PesquisaController;
 use App\Http\Controllers\Produtos\ProdutoController;
-use App\Http\Controllers\SinodalController;
+use App\Http\Controllers\Instancias\SinodalController;
 use App\Http\Controllers\TutorialController;
 use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Auth;
